Add Circular Geofences and DTOs for Controllers.
Add the capability to define Geofences using a center and  a radius as well as the existing polygon approach.
Add DTOs for the web project controllers to use that are separate from the business models. Necessary for backwards compatibility and maintaining a versioned interface.
Handle EntityValidation exceptions from the DB as 400s.